refactor: change type Settings to PasswordSettings in ObnoardingForm and saveOnboardingInfo and rename SettingPasswordForm to UpdatePasswordForm
