+++ src/lib/request/saveOnboardingInfo.ts
@@ -1,7 +1,9 @@
 import { api } from "@/lib/request/api";
-import type { Settings } from "@/types/settings";
+import type { PasswordSettings } from "@/types/settings";
 
-export async function saveOnboardingInfo(payload: Settings): Promise<void> {
+export async function saveOnboardingInfo(
+  payload: PasswordSettings,
+): Promise<void> {
   return api("/api/onboarding", {
     method: "POST",
     body: JSON.stringify(payload),
