+++ src/pages/setting/SettingSSH.tsx
@@ -1,12 +1,12 @@
 import SettingLinuxUsernameForm from "@/components/setting/SettingLinuxUsernameForm";
-import SettingPasswordForm from "@/components/setting/SettingPasswordForm";
+import UpdatePasswordForm from "@/components/setting/UpdatePasswordForm";
 import SettingKeyTable from "@/components/setting/SettingKeyTable";
 
 export default function SettingSSH() {
   return (
     <div className="w-full min-w-lg max-w-2xl px-6 py-8 flex flex-col gap-10">
       <SettingLinuxUsernameForm />
-      <SettingPasswordForm />
+      <UpdatePasswordForm />
       <SettingKeyTable />
     </div>
   );
